at about 14-15 seconds now...

+++ RSALib/BigNum.h
@@ -45,8 +45,8 @@ public:
 	BigNum karatsubaMultiply( BigNum& other );
 	BigNum modInverse( BigNum& other );
 
-	void left_shift( ULONGLONG numShifts, BigNum* retVal );
-	void right_shift( ULONGLONG numShifts, BigNum* retVal );
+	void left_shift( DWORD numShifts, BigNum* retVal );
+	void right_shift( DWORD numShifts, BigNum* retVal );
 
 	void clear();
 	bool toArray(char* array, DWORD& len);
